retrieval/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
numpy==2.2.3
faiss-cpu==1.10.0
httpx==0.28.1
